mdss_dsi_panel: added function declaration of lazyplug_enter_lazy
One of the cherry-picked commits didn't declare the function lazyplug_enter_lazy (in the file or within one of the header files), so the declaration had to be added manually.

The declaration was originally added just prior to the function in which lazyplug_enter_lazy() is being called, however it turns out the version of C I'm using, requires all the variables definitions and declarations to be made at the beginning. This lead to the following error while compiling: warning: ISO C90 forbids mixed declarations and code [-Wdeclaration-after-statement]
More info on the same can be found in this Stackoverflow answer: https://stackoverflow.com/questions/42262965/warning-iso-c90-forbids-mixing-declarations-and-code-wdeclaration-after-state

I need to upgrade to a newer version of GCC ASAP